pyproject.toml
[project]
name = "journal"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url==2.2.0",
    "dj-stripe==2.8.4",
    "django-allauth==0.63.6",
    "django-anymail[sendgrid]==11.0.1",
    "django-debug-toolbar==4.4.6",
    "django-denied==1.3",
    "django-extensions==3.2.3",
    "django-hashid-field==3.4.1",
    "django-simple-history==3.7.0",
    "django-waffle==4.1.0",
    "django>=5.1",
    "environs>=11.2.1",
    "gunicorn==22.0.0",
    "huey>=2.5.2",
    "python-dateutil==2.9.0.post0",
    "sentry-sdk==2.9.0",
    "whitenoise[brotli]==6.7.0",
    "django-sesame>=3.2.2",
]

[dependency-groups]
dev = [
    "factory-boy==3.3.0",
    "honcho>=2.0.0",
    "pre-commit>=4.0.1",
    "pytest-cov==5.0.0",
    "pytest-django>=4.9.0",
    "pytest-socket==0.7.0",
    "pytest-xdist==3.6.1",
    "ruff==0.5.1",
    "time-machine==2.14.2",
]

[tool.coverage.report]
fail_under = 100
omit = [
    "*/migrations/*",
]
show_missing = true
skip_covered = true

# I should replace this with ruff exclusively after updating my Neovim config.
[tool.isort]
profile = "black"
src_paths = ["journal"]

[tool.pytest.ini_options]
addopts = "--nomigrations --disable-socket"
DJANGO_SETTINGS_MODULE = "project.testing_settings"
filterwarnings = [
    # Enable warnings as errors.
    "error",
    # This warning is coming from dj-stripe 2.8.3.
    "ignore:.*The default scheme will be changed.*:PendingDeprecationWarning",
]
# This warning is coming from dj-stripe 2.5.1 (Example)
#    ignore:.*The providing_args argument is deprecated\..*:DeprecationWarning
#    ignore:.*The providing_args argument is deprecated\..*:PendingDeprecationWarning
testpaths = [
    "journal",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    # bandit: Use of `assert` detected
    "S101",
]
